Servicios turisticos con las lecturas con DOM de geonames

+++ estaciones-PeronaCarrillo/src/main/java/servicio/IServicioSitiosTuristicos.java
@@ -2,12 +2,14 @@ package servicio;
 
 import java.util.List;
 
+import org.w3c.dom.NodeList;
+
 import dominio.SitioTuristico;
 
 public interface IServicioSitiosTuristicos {
 
 	
-	List<SitioTuristico> getSitiosInteres(double cordX, double cordY);
+	List<SitioTuristico> getSitiosInteres(int cordX, int cordY);
 	
 	String getInfoSitio(String id);
 }
